src/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-firestore==2.13.1
google-cloud-speech==2.22.0
google-cloud-storage==2.13.0
google-generativeai==0.3.1
python-multipart==0.0.6
pydantic==2.5.2
python-dotenv==1.0.0
aiofiles==23.2.1
Pillow==10.1.0
numpy==1.26.2
asyncpg==0.29.0
psycopg2-binary==2.9.9
google-cloud-sql-python-connector==1.4.3